requirements.txt
llama-index
chromadb
streamlit
langchain
pypdf
langchain-community
langchain_groq